Factor out pin_description.rs
