+++ src/hw/mod.rs
@@ -4,9 +4,9 @@ use std::fmt::{Display, Formatter};
 use std::fs::File;
 use std::io;
 use std::io::{BufReader, Write};
-use std::slice::Iter;
 
 use chrono::{DateTime, Utc};
+use pin_description::PinDescriptionSet;
 use serde::{Deserialize, Serialize};
 
 use crate::hw::pin_descriptions::*;
@@ -18,6 +18,7 @@ use crate::hw::pin_function::PinFunction;
 #[cfg_attr(feature = "pi", path = "pi_hw.rs")]
 #[cfg_attr(not(feature = "pi"), path = "fake_hw.rs")]
 mod implementation;
+pub(crate) mod pin_description;
 mod pin_descriptions;
 pub(crate) mod pin_function;
 
@@ -43,11 +44,13 @@ pub fn get() -> impl Hardware {
     implementation::get()
 }
 
+/// [HardwareDetails] captures a number of specific details about the Hardware we are connected to
 #[derive(Clone, Debug)]
 pub struct HardwareDetails {
     pub hardware: String,
     pub revision: String,
     pub serial: String,
+    /// A Human friendly Hardware Model description
     pub model: String,
 }
 
@@ -147,77 +150,6 @@ impl Display for InputPull {
             InputPull::PullDown => write!(f, "Pull Down"),
             InputPull::None => write!(f, "None"),
         }
-    }
-}
-
-/// [board_pin_number] refer to the pins by the number of the pin printed on the board
-/// [bcm_pin_number] refer to the pins by the "Broadcom SOC channel" number,
-/// these are the numbers after "GPIO"
-#[derive(Debug, Clone)]
-pub struct PinDescription {
-    pub board_pin_number: BoardPinNumber,
-    pub bcm_pin_number: Option<BCMPinNumber>,
-    pub name: &'static str,
-    pub options: &'static [PinFunction], // The set of functions the pin can have, chosen by user config
-}
-
-/// Struct describing all the pins for the connected hardware.
-/// Array indexed from 0 so, board_pin_number -1, as pin numbering start at 1
-#[derive(Debug, Clone)]
-pub struct PinDescriptionSet {
-    pins: [PinDescription; 40],
-}
-
-/// `PinDescriptionSet` describes a set of Pins on a device, using `PinDescription`s
-impl PinDescriptionSet {
-    /// Create a new PinDescriptionSet, from a const array of PinDescriptions
-    pub const fn new(pins: [PinDescription; 40]) -> PinDescriptionSet {
-        PinDescriptionSet { pins }
-    }
-
-    pub fn iter(&self) -> Iter<PinDescription> {
-        self.pins.iter()
-    }
-
-    /// Find a possible pin's board_pin_number using a BCMPinNumber
-    pub fn bcm_to_board(&self, bcm_pin_number: BCMPinNumber) -> Option<BoardPinNumber> {
-        for pin in &self.pins {
-            if pin.bcm_pin_number == Some(bcm_pin_number) {
-                return Some(pin.board_pin_number);
-            }
-        }
-        None
-    }
-
-    /// Return a slice of PinDescriptions
-    pub fn pins(&self) -> &[PinDescription] {
-        &self.pins
-    }
-
-    /// Return a set of PinDescriptions *only** for pins that have BCM pin numbering
-    pub fn bcm_pins(&self) -> Vec<&PinDescription> {
-        self.pins
-            .iter()
-            .filter(|pin| pin.options.len() > 1)
-            .filter(|pin| pin.bcm_pin_number.is_some())
-            .collect::<Vec<&PinDescription>>()
-    }
-
-    /// Return a set of PinDescriptions *only** for pins that have BCM pin numbering, sorted in
-    /// ascending order of [BCMPinNumber]
-    pub fn bcm_pins_sorted(&self) -> Vec<&PinDescription> {
-        let mut pins = self.bcm_pins();
-        pins.sort_by_key(|pin| pin.bcm_pin_number.unwrap());
-        pins
-    }
-}
-
-impl Display for PinDescription {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        writeln!(f, "Board Pin #: {}", self.board_pin_number)?;
-        writeln!(f, "\tBCM Pin #: {:?}", self.bcm_pin_number)?;
-        writeln!(f, "\tName Pin #: {}", self.name)?;
-        writeln!(f, "\tFunctions #: {:?}", self.options)
     }
 }
