+++ src/hw/pin_descriptions.rs
@@ -1,4 +1,5 @@
-use crate::hw::{InputPull, PinDescription, PinFunction};
+use crate::hw::pin_description::PinDescription;
+use crate::hw::{InputPull, PinFunction};
 
 /// This module codifies the descriptions if the Raspberry Pi GPIO hardware
 /// exposed pins, including multiple options (functions) available for some pins
